pyproject.toml
[project]
name = "agentic-ai-sandbox"
version = "0.1.0"
description = "Example implementations of Agentic AI concepts using LangChain and related libraries."
requires-python = ">=3.13"
dependencies = [
    "agentevals>=0.0.9",
    "google-api-python-client>=2.182.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.1",
    "langchain-anthropic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.3",
    "langchain-google-genai>=3.1.0",
    "langchain-google-vertexai>=3.0.3",
    "langchain-mcp-adapters>=0.1.13",
    "langchain-model-profiles>=0.0.4",
    "langchain-openai>=1.1.1",
    "langchain-tavily>=0.2.13",
    "langchain-text-splitters>=1.0.0",
    "langchain[anthropic,openai]>=1.1.3",
    "langgraph>=1.0.3",
    "langgraph-api>=0.5.37",
    "langgraph-cli>=0.4.9",
    "langgraph-swarm>=0.0.14",
    "langmem>=0.0.29",
    "langsmith>=0.4.43",
    "mcp>=1.21.1",
    "pydantic>=2.11.9",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.0",
    "scipy>=1.16.3",
    "sounddevice>=0.5.3",
    "tavily>=1.1.0",
    "tavily-python>=0.7.13",
]
